phd-game-master/phd_game_master/requirements-assets.txt
crewai[tools]      # brings DALL‑E + file helpers
python-dotenv      # for .env loading
requests           # simple HTTP for Freesound / ConfyUI
Pillow             # image post‑processing